requirements.txt
requests
requests[socks]
python-telegram-bot
bs4
user_agent
ping3
python-nmap
tcping
pyping
youtube_dl
mega.py
requests_toolbelt
telethon
tenacity
cryptocode